DTS-32265-Renamed the index to support AWS document Db to improve the rollout process.
